+++ core/createText.js
